requirements.txt
numpy==1.18.0
pandas==0.25.3
opencv-python==4.1.2.30
dlib==19.7.0
Flask==1.1.1
altair==4.0.0
requests==2.22.0
scipy==1.4.1
imutils==0.5.3
tensorflow==1.15
opencv-contrib-python==4.1.2.30
autopy==3.1.0
matplotlib==3.1.2
cmake
seaborn==0.10.0
flask-cors-3.0.8
tensorflow==2.0.0
time
re
os
wtforms
flask_sqlalchemy
datetime
uuid
utils
jinja2
collections
watchdog
flask_bootstrap
flask_wtf
flask_login
cryptography
simplejson
kafka
threading
pathlib
csv
autopy
seaborn
imutils


